Pull out MetadataAgent into a separate file.
