+++ src/Makefile
@@ -41,6 +41,7 @@ endif
 
 SRCS=\
      metadatad.cc \
+     agent.cc \
      api_server.cc \
      store.cc \
      reporter.cc \
