+++ src/agent.cc
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2018 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+
+#include "agent.h"
+
+#include "api_server.h"
+
+namespace google {
+
+MetadataAgent::MetadataAgent(const MetadataAgentConfiguration& config)
+    : config_(config), store_(config_) {}
+
+MetadataAgent::~MetadataAgent() {}
+
+void MetadataAgent::start() {
+  metadata_api_server_.reset(new MetadataApiServer(
+      *this, config_.MetadataApiNumThreads(), "0.0.0.0",
+      config_.MetadataApiPort()));
+  reporter_.reset(new MetadataReporter(
+      this, config_.MetadataReporterIntervalSeconds()));
+}
+
+}
