+++ src/metadatad.cc
@@ -16,7 +16,7 @@
 
 //#include "config.h"
 
-#include "api_server.h"
+#include "agent.h"
 #include "configuration.h"
 #include "docker.h"
 #include "instance.h"
