+++ src/updater.h
@@ -24,7 +24,7 @@
 #include <thread>
 #include <vector>
 
-#include "api_server.h"
+#include "agent.h"
 #include "resource.h"
 #include "time.h"
